EditDialog: refactoring - remove CurrentContext
