+++ src/components/FeaturePanel/EditDialog/EditContent/FeatureEditSection/ConvertNodeToRelation.tsx
@@ -1,5 +1,4 @@
-import { useEditContext } from '../../EditContext';
-import { useCurrentItem } from './CurrentContext';
+import { useCurrentItem, useEditContext } from '../../EditContext';
 import { Alert, Button } from '@mui/material';
 import { NwrIcon } from '../../../NwrIcon';
 import { t } from '../../../../../services/intl';
