+++ src/components/FeaturePanel/EditDialog/EditContent/FeatureEditSection/OpeningHoursEditor/OpeningHoursInput.tsx
@@ -3,7 +3,7 @@ import { TextField } from '@mui/material';
 import { t, Translation } from '../../../../../../services/intl';
 import { encodeUrl } from '../../../../../../helpers/utils';
 import React from 'react';
-import { useCurrentItem } from '../CurrentContext';
+import { useCurrentItem } from '../../../EditContext';
 
 const CantEditText = () => {
   const { tags } = useCurrentItem();
